backend/requirements.txt
# Django & DRF
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Supabase
supabase==2.4.2
postgrest==0.16.8

# Auth / JWT
PyJWT==2.8.0
cryptography==42.0.5

# Database (Supabase PostgreSQL via pooler)
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# AI / ML
openai==1.30.1
tensorflow==2.20.0
numpy>=1.26.4
scikit-learn==1.6.1

# Storage / Utils
python-dotenv==1.0.1
requests==2.31.0
Pillow==11.2.1
python-decouple==3.8
python-whois==0.9.4
feedparser==6.0.11

# Server
gunicorn==22.0.0
whitenoise==6.6.0

# Redis / Rate-limiting
redis==5.0.4
django-ratelimit==4.1.0

# Security
django-axes==6.4.0

# Testing
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0
